moved main.rs into its own binary called src/bin/port_scanner
