+++ engine/src/bin/port_scanner/cli/input/combine.rs
@@ -1,6 +1,6 @@
 use cidr_utils::cidr::{IpCidrIpAddrIterator, IpCidr};
 use std::net::{IpAddr, SocketAddr};
-use crate::model::PortInput;
+use netcore::model::PortInput;
 use super::parser::AddressInput;
 
 #[derive(Debug)]
