+++ engine/src/bin/port_scanner/cli/input/parser.rs
@@ -4,7 +4,7 @@ use std::{
 };
 
 use serde::Serialize;
-use crate::model::PortInput;
+use netcore::model::PortInput;
 use cidr_utils::cidr::IpCidr;
 use crate::cli::error::Error;
 
@@ -39,37 +39,6 @@ pub fn address_parser(src: &str) -> Result<AddressInput, Error> {
             Err(_) => Ok(AddressInput::Pair(src.parse()?))
         }
     }
-}
-
-
-
-impl std::str::FromStr for PortInput {
-    type Err = Error;
-
-    /// Accepts '80-443', '80', '0-10'
-    fn from_str(src: &str) -> Result<Self, Self::Err> {
-        port_parser(src)
-    }
-}
-
-
-pub fn port_parser(src: &str) -> Result<PortInput, Error> {
-    let data: Vec<&str> = src.split("-").collect();
-    let mut bottom = data.get(0).unwrap().parse::<u16>()?;
-
-    if data.len() > 2 {
-        return Err(Error::CliError("port ranges (-p 20-25) can only be in groups of 2 (not 20-25-30)".to_string()))
-    }
-    
-    if data.len() == 2 {
-        let mut top = data.get(1).unwrap().parse::<u16>()?;
-        if bottom >= top {
-            std::mem::swap(&mut bottom, &mut top);
-        }
-        return Ok(PortInput::Range(bottom..top))
-    }
-
-    Ok(PortInput::Singleton(bottom))
 }
