+++ engine/src/lib.rs
@@ -6,3 +6,4 @@ pub mod model;
 pub mod pool;
 pub mod util;
 pub mod netlib;
+pub mod lua;
